+++ src/main/java/com/skd/carrymechanics/mixin/PlayerRenderStateMixin.java
@@ -25,4 +25,14 @@ public class PlayerRenderStateMixin implements ICarryOnRenderState {
     public float carry_mechanics$getRenderWidth() { return carry_mechanics$width; }
     @Override
     public void carry_mechanics$setRenderWidth(float w) { this.carry_mechanics$width = w; }
+
+    @Mixin(AvatarRenderer.class)
+    public static class AvatarExtractorMixin {
+        @Inject(method = "extractRenderState",
+                at = @At("TAIL"))
+        private void onExtractRenderState(AbstractClientPlayer entity, AvatarRenderState state, float partialTicks, CallbackInfo ci) {
+            var data = CarryDataManager.getCarryData(entity);
+            ((ICarryOnRenderState) state).carry_mechanics$setCarryData(data);
+        }
+    }
 }